# Back-End/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Desafio Share Energy API",
    "description": "This is an API Rent",
    "version": "1.0.0",
    "contact": {
      "name": "Saagas Code",
      "email": "saagas.code@gmail.com",
      "url": "https://portfolio-saagas-code.vercel.app/"
    }
  },
  "paths": {
...
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create a user",
        "description": "Create a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "fullname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "example": {
                  "username": "saagas",
                  "fullname": "Matthew Oliver",
                  "email": "saagas.code@gmail.com",
                  "password": "hello8819"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "fullname": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "isAdmin": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "clients": {
                        "type": "array"
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email/Username Already Exists"
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "List all users if you are a admin",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Param for Search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "fullname": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "isAdmin": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "clients": {
                        "type": "array"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "List one users",
        "description": "List one users by _id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": "true",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "fullname": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "isAdmin": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "clients": {
                        "type": "array"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User Not FOUND"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete one user",
        "description": "List one users by _id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": "true",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "User Not FOUND"
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update a user",
        "description": "Update a user, you need to be admin or your id in path",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": "true",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "fullname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "example": {
                  "username": "saagas",
                  "fullname": "Matthew Oliver",
                  "email": "saagas.code@gmail.com",
                  "password": "hello8819"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "fullname": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "isAdmin": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "clients": {
                        "type": "array"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}